// frontend/next-app/src/types/book_type.tsx
type BookType = {
    title: string;
    authors: string[];
    isbn13: string;
    isbn10: string; 
    page_count: number;
    published_date: string;
    categories: string[];
    average_rating: number;
    image_url: string;
    description: string;
    source: string;
};

export default BookType;